testtootta

+++ utilities/ft_putchar.c
@@ -6,13 +6,13 @@
 /*   By: kuyamagi < kuyamagi@student.42tokyo.jp>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/07 15:02:11 by kuyamagi          #+#    #+#             */
-/*   Updated: 2025/04/11 20:52:26 by kuyamagi         ###   ########.fr       */
+/*   Updated: 2025/04/12 18:33:26 by kuyamagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 
-ssize_t	ft_putchar(int c)
+int	ft_putchar(int c)
 {
 	return (write(1, &c, 1));
 }
